Launcher now the same as other projects of mine.
The launcher is now in sync with other projects of mine. This new launcher might be a bit overkill (it has options like finding a custom `python.exe` to run python scripts on) and supports way more than just `.bat` and searches through more directories. But, if we only use a subset of the functionality, we only trigger a subset of the functionality. So it is 100% fine.